+++ lib/src/widgets/raw_editor/raw_editor_actions.dart
@@ -76,7 +76,8 @@ class QuillEditorDeleteTextAction<T extends DirectionalTextEditingIntent>
 
   @override
   bool get isActionEnabled =>
-      !state.widget.readOnly && state.textEditingValue.selection.isValid;
+      !state.widget.configurations.isReadOnly &&
+      state.textEditingValue.selection.isValid;
 }
 
 class QuillEditorUpdateTextSelectionAction<
@@ -93,8 +94,8 @@ class QuillEditorUpdateTextSelectionAction<
     final selection = state.textEditingValue.selection;
     assert(selection.isValid);
 
-    final collapseSelection =
-        intent.collapseSelection || !state.widget.selectionEnabled;
+    final collapseSelection = intent.collapseSelection ||
+        !state.widget.configurations.selectionEnabled;
     // Collapse to the logical start/end.
     TextSelection collapse(TextSelection selection) {
       assert(selection.isValid);
@@ -217,7 +218,8 @@ class QuillEditorExtendSelectionOrCaretPositionAction extends ContextAction<
 
   @override
   bool get isActionEnabled =>
-      state.widget.selectionEnabled && state.textEditingValue.selection.isValid;
+      state.widget.configurations.selectionEnabled &&
+      state.textEditingValue.selection.isValid;
 }
 
 class QuillEditorUpdateTextSelectionToAdjacentLineAction<
@@ -251,8 +253,8 @@ class QuillEditorUpdateTextSelectionToAdjacentLineAction<
   void invoke(T intent, [BuildContext? context]) {
     assert(state.textEditingValue.selection.isValid);
 
-    final collapseSelection =
-        intent.collapseSelection || !state.widget.selectionEnabled;
+    final collapseSelection = intent.collapseSelection ||
+        !state.widget.configurations.selectionEnabled;
     final value = state.textEditingValue;
     if (!value.selection.isValid) {
       return;
@@ -307,7 +309,7 @@ class QuillEditorSelectAllAction extends ContextAction<SelectAllTextIntent> {
   }
 
   @override
-  bool get isActionEnabled => state.widget.selectionEnabled;
+  bool get isActionEnabled => state.widget.configurations.selectionEnabled;
 }
 
 class QuillEditorCopySelectionAction
@@ -559,7 +561,7 @@ class QuillEditorApplyLinkAction extends Action<QuillEditorApplyLinkIntent> {
         return LinkStyleDialog(
           text: initialTextLink.text,
           link: initialTextLink.link,
-          dialogTheme: state.widget.dialogTheme,
+          dialogTheme: state.widget.configurations.dialogTheme,
         );
       },
     );